GMT_DOS-M1-CTRL examples/step -> examples/loadcells
 examples/loadcells implement both loadcells model:
rust-based and fem-based
